\name{Crosstabs.Loglinear-package}

\alias{Crosstabs.Loglinear-package}

\docType{package}

\title{Crosstabs.Loglinear}

\description{Provides 'SPSS'- and 'SAS'-like output for cross tabulations of two
	categorical variables (CROSSTABS) and for hierarchical loglinear analyses of two or more
	categorical variables (LOGLINEAR).
}

\references{
	{Agresti, A. (2013). \emph{Categorical data analysis (3rd ed).} Hobokon, NJ: John Wiley & Sons.}
\cr\cr {Ajzen, R., & Walker, C. M. (2021). \emph{Categorical data analysis for the behavioral and 
	    social sciences (2nd ed.).} New York, NY: Routledge.}
\cr\cr {Field, A. (2018). Chapter 18: Categorical data. 
		\emph{Discovering statistics using SPSS (5th ed.).} Los Angeles, CA: Sage.}
\cr\cr {Noursis, M. J. (2012). Chapter 1: Model selection loglinear analysis.
	\emph{IBM SPSS statistics 19: Advanced statistical 
	procedures Companion.} Upper Saddle River, NJ: Prentice Hall.}
\cr\cr {Nussbaum, E. M. (2015). \emph{Categorical and nonparametric data analysis 
		choosing the best statistical technique.} New York, NY: Routledge.}
\cr\cr {Stevens, J. P. (2009). Chapter 14: Categorical data analysis: The log linear model. 
	\emph{Applied multivariate statistics for the social sciences (5th ed.).} 
	New York, NY: Routledge.}
\cr\cr {Tabachnick, B. G., & Fidell, L. S. (2019). Chapter 16: Multiway 
	frequency analysis. \emph{Using multivariate statistics.} New York, NY: Pearson.}
\cr\cr {von Eye, A., & Mun, E. Y. (2013). \emph{Log-Linear modeling concepts, 
		interpretation, and application.} Hoboken, NJ: Wiley.}
}



